🔨Fix test that checks if colgroup members are unique
When verifying that all column groups are unique, ensure that no two
colgroups have the same set of casefolded member column names. This
ensures that I do not accidentally specify a column group definition
that "shadows" the other, e.g.:

  __Size = ["SizeX", "SizeY"]
  __SizeXY = ["sizex", "sizey"]

Here, "__Size" will shadow "__SizeXY" because both column group
definitions share the same set of casefolded column names (which is
{"sizex", "sizey"}).

Note that a column group can still shadow another if it is a strict
superset:

  __SizeXYZ = ["SizeX", "SizeY", "SizeZ"]
  __SizeXY = ["sizex", "sizey"]

Here, "__SizeXYZ" will shadow "SizeXY" if the TXT file has the columns
"SizeX", "SizeY", "SizeZ". This is an intentional design choice.